+++ fedot/core/optimisers/gp_comp/gp_optimiser.py
@@ -1,32 +1,21 @@
-import logging
 from copy import deepcopy
 from random import choice
-from typing import Any, List, Optional, Sequence, Union
-
-from tqdm import tqdm
+from typing import Any, List, Optional, Sequence, Union, Callable
 
 from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
 from fedot.core.constants import MAXIMAL_ATTEMPTS_NUMBER
-from fedot.core.optimisers.archive import GenerationKeeper
-from fedot.core.optimisers.gp_comp.evaluation import MultiprocessingDispatcher
 from fedot.core.optimisers.gp_comp.individual import Individual
 from fedot.core.optimisers.gp_comp.operators.crossover import CrossoverTypesEnum, crossover
 from fedot.core.optimisers.gp_comp.operators.elitism import Elitism, ElitismTypesEnum
 from fedot.core.optimisers.gp_comp.operators.inheritance import GeneticSchemeTypesEnum, inheritance
-from fedot.core.optimisers.gp_comp.operators.mutation import MutationTypesEnum, Mutation
+from fedot.core.optimisers.gp_comp.operators.mutation import MutationTypesEnum
 from fedot.core.optimisers.gp_comp.operators.operator import PopulationT
 from fedot.core.optimisers.gp_comp.operators.regularization import RegularizationTypesEnum, regularized_population
 from fedot.core.optimisers.gp_comp.operators.selection import SelectionTypesEnum, crossover_parents_selection, selection
-from fedot.core.optimisers.gp_comp.parameters.graph_depth import AdaptiveGraphDepth
-from fedot.core.optimisers.gp_comp.parameters.operators_prob import init_adaptive_operators_prob
-from fedot.core.optimisers.gp_comp.parameters.population_size import PopulationSize, init_adaptive_pop_size
-from fedot.core.optimisers.graph import OptGraph
-from fedot.core.optimisers.objective import GraphFunction, ObjectiveFunction
+from fedot.core.optimisers.gp_comp.populational_optimiser import PopulationalOptimiser
 from fedot.core.optimisers.objective.objective import Objective
-from fedot.core.optimisers.optimizer import GraphGenerationParams, GraphOptimiser, GraphOptimiserParameters
-from fedot.core.optimisers.timer import OptimisationTimer
+from fedot.core.optimisers.optimizer import GraphGenerationParams, GraphOptimiserParameters
 from fedot.core.pipelines.pipeline import Pipeline
-from fedot.core.utilities.grouped_condition import GroupedCondition
 
 
 class GPGraphOptimiserParameters(GraphOptimiserParameters):
@@ -82,7 +71,7 @@ class GPGraphOptimiserParameters(GraphOptimiserParameters):
         self.set_default_params()  # always initialize in proper state
 
 
-class EvoGraphOptimiser(GraphOptimiser):
+class EvoGraphOptimiser(PopulationalOptimiser):
     """
     Multi-objective evolutionary graph optimiser named GPComp
     """
@@ -94,59 +83,7 @@ class EvoGraphOptimiser(GraphOptimiser):
                  graph_generation_params: GraphGenerationParams,
                  parameters: Optional[GPGraphOptimiserParameters] = None):
         super().__init__(objective, initial_graphs, requirements, graph_generation_params, parameters)
-        self.parameters = parameters or GPGraphOptimiserParameters()
         self.elitism = Elitism(self.parameters.elitism_type, requirements, objective.is_multi_objective)
-        self.mutation = Mutation(parameters.mutation_types, graph_generation_params, requirements)
-        self.population = None
-        self.generations = GenerationKeeper(self.objective, keep_n_best=requirements.keep_n_best)
-        self.timer = OptimisationTimer(timeout=self.requirements.timeout)
-        self.eval_dispatcher = MultiprocessingDispatcher(graph_adapter=graph_generation_params.adapter,
-                                                         timer=self.timer,
-                                                         n_jobs=requirements.n_jobs,
-                                                         graph_cleanup_fn=_unfit_pipeline)
-
-        # stopping_after_n_generation may be None, so use some obvious max number
-        max_stagnation_length = parameters.stopping_after_n_generation or requirements.num_of_generations
-        self.stop_optimisation = \
-            GroupedCondition().add_condition(
-                lambda: self.timer.is_time_limit_reached(self.current_generation_num),
-                'Optimisation stopped: Time limit is reached'
-            ).add_condition(
-                lambda: self.current_generation_num >= requirements.num_of_generations + 1,
-                'Optimisation stopped: Max number of generations reached'
-            ).add_condition(
-                lambda: self.generations.stagnation_duration >= max_stagnation_length,
-                'Optimisation finished: Early stopping criteria was satisfied'
-            )
-
-        # Define adaptive parameters
-
-        self._pop_size: PopulationSize = \
-            init_adaptive_pop_size(parameters.genetic_scheme_type, requirements, self.generations)
-
-        self._operators_prob = \
-            init_adaptive_operators_prob(parameters.genetic_scheme_type, requirements)
-
-        start_depth = requirements.start_depth or requirements.max_depth
-
-        self._graph_depth = AdaptiveGraphDepth(self.generations,
-                                               start_depth=start_depth,
-                                               max_depth=requirements.max_depth,
-                                               max_stagnated_generations=parameters.depth_increase_step,
-                                               adaptive=parameters.with_auto_depth_configuration)
-
-        # Define parameters
-
-        self.requirements.max_depth = self._graph_depth.initial
-
-        self.requirements.pop_size = self._pop_size.initial
-
-        self.initial_individuals = \
-            [Individual(self.graph_generation_params.adapter.adapt(graph)) for graph in initial_graphs]
-
-    @property
-    def current_generation_num(self) -> int:
-        return self.generations.generation_num
 
     def _extend_population(self, initial_individuals):
         iter_num = 0
@@ -169,78 +106,39 @@ class EvoGraphOptimiser(GraphOptimiser):
         self.mutation.update_requirements(self.requirements)
         return initial_individuals
 
-    def _next_population(self, next_population: PopulationT):
-        self.update_native_generation_numbers(next_population)
-        self.generations.append(next_population)
-        self._optimisation_callback(next_population, self.generations)
-        self.population = next_population
+    def _evolution_process(self, evaluator: Callable):
+        """ Method realizing the complete process of evolution """
+        # Adding of initial assumptions to history as zero generation
+        self._next_population(evaluator(self.initial_individuals))
 
-        self.log.info(f'Generation num: {self.current_generation_num}')
-        self.log.info(f'Best individuals: {str(self.generations)}')
-        self.log.info(f'no improvements for {self.generations.stagnation_duration} iterations')
-        self.log.info(f'spent time: {round(self.timer.minutes_from_start, 1)} min')
-
-    def set_evaluation_callback(self, callback: Optional[GraphFunction]):
-        # Redirect callback to evaluation dispatcher
-        self.eval_dispatcher.set_evaluation_callback(callback)
-
-    def update_native_generation_numbers(self, population: PopulationT):
-        for individual in population:
-            individual.set_native_generation(self.current_generation_num)
-
-    def optimise(self, objective: ObjectiveFunction,
-                 show_progress: bool = True) -> Sequence[OptGraph]:
-
-        # eval_dispatcher defines how to evaluate objective on the whole population
-        evaluator = self.eval_dispatcher.dispatch(objective)
-
-        with self.timer, tqdm(total=self.requirements.num_of_generations,
-                              desc='Generations', unit='gen', initial=1,
-                              disable=not show_progress or self.log.logging_level == logging.NOTSET):
-
-            # Adding of initial assumptions to history as zero generation
+        if len(self.initial_individuals) < self.requirements.pop_size:
+            self.initial_individuals = self._extend_population(self.initial_individuals)
+            # Adding of extended population to history
             self._next_population(evaluator(self.initial_individuals))
 
-            if len(self.initial_individuals) < self.requirements.pop_size:
-                self.initial_individuals = self._extend_population(self.initial_individuals)
-                # Adding of extended population to history
-                self._next_population(evaluator(self.initial_individuals))
+        while not self.stop_optimisation():
+            self._update_requirements()
+            individuals_to_select = regularized_population(self.parameters.regularization_type,
+                                                           self.population,
+                                                           evaluator,
+                                                           self.graph_generation_params)
 
-            while not self.stop_optimisation():
-                self._update_requirements()
-                individuals_to_select = regularized_population(self.parameters.regularization_type,
-                                                               self.population,
-                                                               evaluator,
-                                                               self.graph_generation_params)
+            selected_individuals = selection(types=self.parameters.selection_types,
+                                             population=individuals_to_select,
+                                             pop_size=self.requirements.pop_size,
+                                             params=self.graph_generation_params)
+            new_population = self._reproduce(selected_individuals)
 
-                selected_individuals = selection(types=self.parameters.selection_types,
-                                                 population=individuals_to_select,
-                                                 pop_size=self.requirements.pop_size,
-                                                 params=self.graph_generation_params)
-                new_population = self._reproduce(selected_individuals)
+            new_population = self.mutation(new_population)
 
-                new_population = self.mutation(new_population)
+            new_population = evaluator(new_population)
 
-                new_population = evaluator(new_population)
+            new_population = self._inheritance(new_population)
 
-                new_population = self._inheritance(new_population)
+            new_population = self.elitism(self.generations.best_individuals, new_population)
 
-                new_population = self.elitism(self.generations.best_individuals, new_population)
-
-                # Adding of new population to history
-                self._next_population(new_population)
-        all_best_graphs = [ind.graph for ind in self.generations.best_individuals]
-        return all_best_graphs
-
-    def _update_requirements(self):
-        if not self.generations.is_any_improved:
-            self.requirements.mutation_prob, self.requirements.crossover_prob = \
-                self._operators_prob.next(self.population)
-        self.requirements.pop_size = self._pop_size.next(self.population)
-        self.requirements.max_depth = self._graph_depth.next()
-        self.log.info(
-            f'Next population size: {self.requirements.pop_size}; max graph depth: {self.requirements.max_depth}')
-        self.mutation.update_requirements(self.requirements)
+            # Adding of new population to history
+            self._next_population(new_population)
 
     def _inheritance(self, offspring: PopulationT) -> PopulationT:
         """Gather next population given new offspring, previous population and elite individuals.
@@ -267,8 +165,3 @@ class EvoGraphOptimiser(GraphOptimiser):
         return crossover(self.parameters.crossover_types, individual1, individual2,
                          max_depth=self.requirements.max_depth, crossover_prob=self.requirements.crossover_prob,
                          params=self.graph_generation_params)
-
-
-def _unfit_pipeline(graph: Any):
-    if isinstance(graph, Pipeline):
-        graph.unfit()
