+++ fedot/core/optimisers/gp_comp/populational_optimiser.py
@@ -1,0 +1,152 @@
+import logging
+from abc import abstractmethod
+from typing import Any, Optional, Sequence
+
+from tqdm import tqdm
+
+from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.archive import GenerationKeeper
+from fedot.core.optimisers.gp_comp.evaluation import MultiprocessingDispatcher
+from fedot.core.optimisers.gp_comp.individual import Individual
+from fedot.core.optimisers.gp_comp.operators.mutation import Mutation
+from fedot.core.optimisers.gp_comp.operators.operator import PopulationT
+from fedot.core.optimisers.gp_comp.parameters.graph_depth import AdaptiveGraphDepth
+from fedot.core.optimisers.gp_comp.parameters.operators_prob import init_adaptive_operators_prob
+from fedot.core.optimisers.gp_comp.parameters.population_size import PopulationSize, init_adaptive_pop_size
+from fedot.core.optimisers.graph import OptGraph
+from fedot.core.optimisers.objective import GraphFunction, ObjectiveFunction
+from fedot.core.optimisers.objective.objective import Objective
+from fedot.core.optimisers.optimizer import GraphGenerationParams, GraphOptimiser
+from fedot.core.optimisers.timer import OptimisationTimer
+from fedot.core.pipelines.pipeline import Pipeline
+from fedot.core.utilities.grouped_condition import GroupedCondition
+
+
+class PopulationalOptimiser(GraphOptimiser):
+    """
+    Base class of populational optimiser.
+    PopulationalOptimiser implements all basic methods for optimisation not related to evolution process
+    to experiment with other kinds of evolution optimization methods
+    It allows to find the optimal solution using specified metric (one or several).
+    To implement the specific evolution strategy,
+    the abstract method '_evolution_process' should be re-defined in the ancestor class
+
+    :param objective: objective for optimisation
+    :param initial_graphs: graphs which were initialized outside the optimiser
+    :param requirements: implementation-independent requirements for graph optimiser
+    :param graph_generation_params: parameters for new graph generation
+    :param parameters: parameters for specific implementation of graph optimiser
+    """
+
+    def __init__(self,
+                 objective: Objective,
+                 initial_graphs: Sequence[Pipeline],
+                 requirements: PipelineComposerRequirements,
+                 graph_generation_params: GraphGenerationParams,
+                 parameters: Optional['GPGraphOptimiserParameters'] = None):
+        super().__init__(objective, initial_graphs, requirements, graph_generation_params, parameters)
+        self.mutation = Mutation(parameters.mutation_types, graph_generation_params, requirements)
+        self.population = None
+        self.generations = GenerationKeeper(self.objective, keep_n_best=requirements.keep_n_best)
+        self.timer = OptimisationTimer(timeout=self.requirements.timeout)
+        self.eval_dispatcher = MultiprocessingDispatcher(graph_adapter=graph_generation_params.adapter,
+                                                         timer=self.timer,
+                                                         n_jobs=requirements.n_jobs,
+                                                         graph_cleanup_fn=_unfit_pipeline)
+
+        # stopping_after_n_generation may be None, so use some obvious max number
+        max_stagnation_length = parameters.stopping_after_n_generation or requirements.num_of_generations
+        self.stop_optimisation = \
+            GroupedCondition().add_condition(
+                lambda: self.timer.is_time_limit_reached(self.current_generation_num),
+                'Optimisation stopped: Time limit is reached'
+            ).add_condition(
+                lambda: self.current_generation_num >= requirements.num_of_generations + 1,
+                'Optimisation stopped: Max number of generations reached'
+            ).add_condition(
+                lambda: self.generations.stagnation_duration >= max_stagnation_length,
+                'Optimisation finished: Early stopping criteria was satisfied'
+            )
+
+        # Define adaptive parameters
+
+        self._pop_size: PopulationSize = \
+            init_adaptive_pop_size(parameters.genetic_scheme_type, requirements, self.generations)
+
+        self._operators_prob = \
+            init_adaptive_operators_prob(parameters.genetic_scheme_type, requirements)
+
+        start_depth = requirements.start_depth or requirements.max_depth
+
+        self._graph_depth = AdaptiveGraphDepth(self.generations,
+                                               start_depth=start_depth,
+                                               max_depth=requirements.max_depth,
+                                               max_stagnated_generations=parameters.depth_increase_step,
+                                               adaptive=parameters.with_auto_depth_configuration)
+
+        # Define parameters
+
+        self.requirements.max_depth = self._graph_depth.initial
+
+        self.requirements.pop_size = self._pop_size.initial
+
+        self.initial_individuals = \
+            [Individual(self.graph_generation_params.adapter.adapt(graph)) for graph in initial_graphs]
+
+    @property
+    def current_generation_num(self) -> int:
+        return self.generations.generation_num
+
+    def _next_population(self, next_population: PopulationT):
+        self.update_native_generation_numbers(next_population)
+        self.generations.append(next_population)
+        self._optimisation_callback(next_population, self.generations)
+        self.population = next_population
+
+        self.log.info(f'Generation num: {self.current_generation_num}')
+        self.log.info(f'Best individuals: {str(self.generations)}')
+        self.log.info(f'no improvements for {self.generations.stagnation_duration} iterations')
+        self.log.info(f'spent time: {round(self.timer.minutes_from_start, 1)} min')
+
+    def set_evaluation_callback(self, callback: Optional[GraphFunction]):
+        # Redirect callback to evaluation dispatcher
+        self.eval_dispatcher.set_evaluation_callback(callback)
+
+    def update_native_generation_numbers(self, population: PopulationT):
+        for individual in population:
+            individual.set_native_generation(self.current_generation_num)
+
+    @abstractmethod
+    def _evolution_process(self, *args, **kwargs):
+        """ Method realizing the complete process of evolution """
+        raise NotImplementedError()
+
+    def optimise(self, objective: ObjectiveFunction,
+                 show_progress: bool = True) -> Sequence[OptGraph]:
+
+        # eval_dispatcher defines how to evaluate objective on the whole population
+        evaluator = self.eval_dispatcher.dispatch(objective)
+
+        with self.timer, tqdm(total=self.requirements.num_of_generations,
+                              desc='Generations', unit='gen', initial=1,
+                              disable=not show_progress or self.log.logging_level == logging.NOTSET):
+
+            self._evolution_process(evaluator=evaluator)
+
+        all_best_graphs = [ind.graph for ind in self.generations.best_individuals]
+        return all_best_graphs
+
+    def _update_requirements(self):
+        if not self.generations.is_any_improved:
+            self.requirements.mutation_prob, self.requirements.crossover_prob = \
+                self._operators_prob.next(self.population)
+        self.requirements.pop_size = self._pop_size.next(self.population)
+        self.requirements.max_depth = self._graph_depth.next()
+        self.log.info(
+            f'Next population size: {self.requirements.pop_size}; max graph depth: {self.requirements.max_depth}')
+        self.mutation.update_requirements(self.requirements)
+
+
+def _unfit_pipeline(graph: Any):
+    if isinstance(graph, Pipeline):
+        graph.unfit()
